requirements.txt
pycryptodome==3.10.1
fastapi
uvicorn
pyrebase4